+++ dms-20250414/alibabacloud_dms20250414/models/_get_data_agent_workspace_info_request.py
@@ -10,9 +10,9 @@ class GetDataAgentWorkspaceInfoRequest(DaraModel):
         dmsunit: str = None,
         workspace_id: str = None,
     ):
-        # The DMS unit.
+        # The current DMS unit.
         self.dmsunit = dmsunit
-        # The workspace ID.
+        # The ID of the collaborative workspace.
         self.workspace_id = workspace_id
 
     def validate(self):