+++ dms-20250414/alibabacloud_dms20250414/models/_list_data_agent_workspace_request.py
@@ -19,25 +19,25 @@ class ListDataAgentWorkspaceRequest(DaraModel):
     ):
         # The current DMS unit.
         self.dmsunit = dmsunit
-        # This parameter is reserved for future use.
+        # The maximum number of entries per page.
         self.max_results = max_results
-        # This parameter is reserved for future use.
+        # The token information.
         self.next_token = next_token
         # The sort order.
         self.order = order
-        # The field to sort by.
+        # The name of the field by which to sort.
         self.order_by = order_by
-        # The page number.
+        # The page number to start from.
         # 
         # This parameter is required.
         self.page_number = page_number
-        # The number of entries to return on each page.
+        # The number of records per page.
         # 
         # This parameter is required.
         self.page_size = page_size
         # The workspace name.
         self.workspace_name = workspace_name
-        # The workspace type.
+        # The type of workspace to query.
         # 
         # This parameter is required.
         self.workspace_type = workspace_type